controllers

+++ app/Http/Controllers/AdminEditController.php
@@ -13,6 +13,9 @@ class AdminEditController extends Controller
 {
     //--------Province--------//
     public function prefecture_new(){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+
         $province = new Province();
         $province->name = '';
         $province->img_url = asset('img/top/test_area/blank_city.png');
@@ -20,32 +23,45 @@ class AdminEditController extends Controller
 
     }
     public function prefecture_view(){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+
         $provinces = Province::all();
         return view('admin.prefecture.view', ['provinces'=>$provinces]);
     }
     public function prefecture_edit($id){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
+        
         $province = Province::find($id);
         return view('admin.prefecture.edit', ['province'=>$province]);
     }
     public function prefecture_upload(Request $request){
-
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $file = $request->file;
         if (strtolower($file->getClientOriginalExtension()) == 'png'
         || strtolower($file->getClientOriginalExtension()) == 'jpg'
         || strtolower($file->getClientOriginalExtension()) == 'jpeg'
         || strtolower($file->getClientOriginalExtension()) == 'gif'
         ) {
-        $filename = date('ymdhis') . '.'.$file->getClientOriginalExtension();
-        $file_url = 'uploads/province/'.$filename;
-        $file->move('uploads/province', $filename);
-        return ['file_url' => $file_url];
+            $filename = date('ymdhis') . '.'.$file->getClientOriginalExtension();
+            $file_url = 'uploads/province/'.$filename;
+            $file->move('uploads/province', $filename);
+            return ['file_url' => $file_url];
         }
         else{
             return 'unsupported image';
         }
     }
-
+    
     public function prefecture_save(Request $request){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
+        
         $province = new Province();
         if($request->province_id > 0) $province = Province::find($request->province_id);
         $province->name = $request->add_test_name;
@@ -53,29 +69,44 @@ class AdminEditController extends Controller
         $province->save();
         return redirect()->route('admin.prefecture.view');
     }
-
+    
     public function prefecture_delete($id){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+
         Province::find($id)->delete();
         return redirect()->route('admin.prefecture.view');
     }
 
     //---------Ganre---------//
     public function ganre_new(){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $ganre = new Ganre();
         $ganre->name = '';
         return view('admin.ganre.edit', ['ganre'=>$ganre]);
 
     }
     public function ganre_view(){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $ganres = Ganre::all();
         return view('admin.ganre.view', ['ganres'=>$ganres]);
     }
     public function ganre_edit($id){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $ganre = Ganre::find($id);
         return view('admin.ganre.edit', ['ganre'=>$ganre]);
     }
 
     public function ganre_save(Request $request){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $ganre = new Ganre();
 
         if($request->ganre_id > 0) $ganre = Ganre::find($request->ganre_id);
@@ -85,12 +116,18 @@ class AdminEditController extends Controller
     }
 
     public function ganre_delete($id){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         Ganre::find($id)->delete();
         return redirect()->route('admin.ganre.view');
     }
 
     //--------Notice--------//
     public function notice_new(){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $notice = new Notice();
         $notice->name = '';
         $notice->notice_img_url = asset('img/top/test_area/blank_city.png');
@@ -98,14 +135,23 @@ class AdminEditController extends Controller
 
     }
     public function notice_view(){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $notices = Notice::all();
         return view('admin.notice.view', ['notices'=>$notices]);
     }
     public function notice_edit($id){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $notice = Notice::find($id);
         return view('admin.notice.edit', ['notice'=>$notice]);
     }
     public function notice_upload(Request $request){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
 
         $file = $request->file;
         if (strtolower($file->getClientOriginalExtension()) == 'png'
@@ -124,6 +170,9 @@ class AdminEditController extends Controller
     }
 
     public function notice_save(Request $request){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $notice = new Notice();
         $date_val = str_replace(['年','月','日'], '', $request->notice_date);
         if($request->notice_id > 0) $notice = Notice::find($request->notice_id);
@@ -136,16 +185,25 @@ class AdminEditController extends Controller
     }
 
     public function notice_delete($id){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         Notice::find($id)->delete();
         return redirect()->route('admin.notice.view');
     }
 
     public function user_view(){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         $users = User::all();
         return view('admin.user.view', ['users'=>$users]);
     }
 
     public function user_delete($id){
+        if(!Auth::check()) return redirect()->route('_login');
+        if(Auth::user()->user_role != 1) return redirect()->route('home');
+        
         User::find($id)->delete();
         return redirect()->route('admin.user.view');
     }
